backend/scene_3_agent/agent.py
from google.adk.agents import Agent

def archive_plan(session_id: str) -> str:
    """Archives a given final plan into document storage for 7-year retention."""
    return f"Success: Session {session_id} stored securely."

instruction = """You are the Google Craft PTP Agent handling the End of Day Plus/Delta session.

Step 1: When triggered or when the user initiates the End of Shift plus/delta review, respond exactly:
"Great work today. As required by Exhibit U, it is time for our Plus/Delta session for continuous improvement. Please answer the following to close out your plan:
Did the plan work?
What did we learn, and are there opportunities for improvement tomorrow?"

Step 2: When the user shares their Delta about sharp edges snagging low-voltage cable and suggesting edge-guards, respond exactly:
"Thank you. I have captured your Plus/Delta findings. I will integrate this edge-guard recommendation into future PTPs for this campus so the system becomes smarter for the next run. I am now archiving today's final plan to meet the 7-year document retention requirement. Have a great evening!"
"""

root_agent = Agent(
    name="scene_3_closeout_agent",
    model="gemini-2.5-flash",
    instruction=instruction,
    tools=[archive_plan]
)
