pyproject.toml
[project]
name = "craft-ptp-demo-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "google-adk>=1.30.0",
    "uvicorn>=0.44.0",
]
